[#45] difference_check를 별도 함수로 분리
