+++ src/action/run/steps/mod.rs
@@ -3,3 +3,6 @@ pub use connection::connect_database;
 
 mod fetching;
 pub use fetching::get_table_list;
+
+mod check;
+pub use check::difference_check;
